+++ R/funs.R
@@ -13,7 +13,7 @@
 #' f <- system.file(package = "fpeek",
 #'   "datafiles", "cigfou-ISO-8859-1.txt")
 #' peek_count_lines(f)
-peek_count_lines <- function(path, with_eof = FALSE ) {
+peek_count_lines <- function(path, with_eof = FALSE) {
   nlines_(path, with_eof = with_eof)
 }
 
@@ -33,8 +33,11 @@ peek_count_lines <- function(path, with_eof = FALSE ) {
 #' peek_head(f, n = 4)
 #' peek_head(f, n = 4, intern = TRUE)
 peek_head <- function(path, n = 10, intern = FALSE) {
-  if( !intern ) head_print_(path, n)
-  else head_str_(path, n)
+  if (!intern) {
+    head_print_(path, n)
+  } else {
+    head_str_(path, n)
+  }
 }
 
 
@@ -49,8 +52,11 @@ peek_head <- function(path, n = 10, intern = FALSE) {
 #' peek_tail(f, n = 4)
 #' peek_tail(f, n = 4, intern = TRUE)
 peek_tail <- function(path, n = 10, intern = FALSE) {
-  if( !intern ) tail_print_(path, n)
-  else tail_str_(path, n)
+  if (!intern) {
+    tail_print_(path, n)
+  } else {
+    tail_str_(path, n)
+  }
 }
 
 
@@ -79,18 +85,14 @@ peek_tail <- function(path, n = 10, intern = FALSE) {
 #' peek_head(newfile, n = 10)
 #'
 peek_iconv <- function(path, from, to = "UTF-8", newfile = NULL) {
-  if( is.null(newfile) ){
-    file_iconv_(path, from, to )
+  if (is.null(newfile)) {
+    file_iconv_(path, from, to)
   } else {
-    file_con <- file( newfile, open = "wt", encoding = to)
+    file_con <- file(newfile, open = "wt", encoding = to)
     sink(file = file_con)
-    file_iconv_(path, from, to )
+    file_iconv_(path, from, to)
     sink()
     close(file_con)
     newfile
   }
-
 }
-
-
-